ai_agent/main.py
import asyncio
from dotenv import load_dotenv
from infrastructure.notion_local_client import notion
from infrastructure.llm_client import openai, client
from ai_agent.domain.models import Task, QuickStartAction, DailyPlan
from ai_agent.agents.task_analyzer import analyze_task_complexity
from ai_agent.agents.motivator import generate_goggins_motivation
from ai_agent.agents.planner import create_daily_plan

load_dotenv()


# TODO: Implement main application logic
def main():
    # Initialize components
    # Run the application
    pass


if __name__ == "__main__":
    main()
