requirements.txt
# Core dependencies
python-dotenv
pydantic
pydantic-settings

# Notion integration
notion-client==2.0.0

# OpenAI and LangChain
openai
langchain
langchain-openai
instructor

# LangGraph
langgraph

# Telegram integration (for future use)
python-telegram-bot==20.7

# Testing
pytest==7.4.3

# Type checking
mypy==1.7.1

# Linting
flake8==6.1.0

# Async support
asyncio==3.4.3

# Date and time handling
python-dateutil==2.8.2